refactor external links to use same component as other menu links
