+++ src/components/ui/page/Navigation/NavigationTooltip.tsx
@@ -1,0 +1,17 @@
+import { Center, Tooltip } from '@chakra-ui/react';
+
+export function NavigationTooltip({ label, children }: { label: string; children: JSX.Element }) {
+  return (
+    <Center>
+      <Tooltip
+        closeDelay={250}
+        gutter={10}
+        hasArrow
+        label={label}
+        placement="right"
+      >
+        {children}
+      </Tooltip>
+    </Center>
+  );
+}
